Record gasmix in plot_data
Now, that we are able to deal with several cylinder pressures
at a given time we lost the notion of "current cylinder". But
we still need the current gas (for heatmap, MOD, etc etc). So
we better store that in the plot_info struct.

Some of that information goes through the diveplotdatamodel
and since the cells cannot hold aggregate types like
struct gasmix, I also added helper functions to pack a
gasmix into a uint32_t.

Fixes #562

Signed-off-by: Robert C. Helling <helling@atdotde.de>