user manual: re-correct pSCR wording
Unsure where and why this got changed in the update to 4.7.

In pSCR world, the gas that is currently driving the rebreater is
called a "driving gas". This is not per definition backgas, as any
gas can be plugged in by means of a swichblock.

Further. The gas that is trown away (typically 10%) is released
from the unit at inhale of the diver. Yes, this may sound weird
to the average OC diver, but it is like that. It's by design.

All this wisdom from a GUE trained RB80 diver :-)

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>