desktop-main.cpp: add validateGL()
Add a new static function to validate the availability of OpenGL
on this particular desktop system. It makes some calls to create
a platform agnostic GL context that renders to a offscreen surface.
Then it makes a couple of glGetIntegerv((GL_xxx_VERSION, ...) calls
to see if the GL profile version is at least 2.1.

In case any of the steps fail, a stderr message is shown and all
QtQuick based widgets would be rendered using a software renderer.

Testing was done in the case of the Google maps plugin and the
fallback seems to work on Windows, but further testing will be
required on all OS.

For the time being, the automatic fallback is only supported
on Qt 5.8.0 or later.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>