desktop-main.cpp: further improve the GL version detection
First, attempt to use glGetString(GL_VERSION), which works
on legacy profiles but has to be parsed (unsafe and vendor specific?).
If the above fails attempt to use the newer version API from
glGetIntegerv().

If both fail, fall back to a software renderer or exit.

If GLES is detected from glGetString() (for some odd reason -
e.g. emulated driver) show a warning and return early as we don't
handle it's versioning.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>