OSTC: initialize CNS correctly
As the OSTC stores the CNS at the start of the dive in the dive header,
we can initialize the CNS value correctly. The current code did
some sort of initialization, and carried CNS data from another dive to
the initial part of the profile (before a CNS sample was received),
resulting in strange data, for example, in the info box.

Notice: requires commit cb38734df61b97 from Suburface-branch of
libdivecomputer

See also: #789

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>