Don't show cylinder pressure graph for wrong dc for merged dives
When merging a logged dive from a dive computer and a planned dive
the cylinders are not always merged. They are e.g. not merged if we
have manually entered pressure values. So we correctly end up with
a doubled set of cylinders.
When now plotting the pressure graphs we need to make sure to only
plot graphs for the cylinders which really belong to the displayed dc.
This commit introduces this behaviour.

Signed-off-by: Stefan Fuchs <sfuchs@gmx.de>